增加拍照存储权限

+++ AKTWaiterCloud/Util/AktUtil.m
@@ -59,7 +59,7 @@
         
         do {
             size_t size = (size_t)(2 * sizeof(short int));
-            read = fread(pcm_buffer, size, PCM_SIZE, pcm);
+            read = (int)fread(pcm_buffer, size, PCM_SIZE, pcm);
             if (read == 0)
                 write = lame_encode_flush(lame, mp3_buffer, MP3_SIZE);
             else
